MAR-2065 Refactored the HederSection component
